@bustos00 [task:Atualizando as camadas]

+++ src/main/java/org/generation/blogPessoal/controller/PostagemController.java
@@ -1,8 +1,11 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
-import org.generation.blogPessoal.model.postagem;
+import javax.validation.Valid;
+
+import org.generation.blogPessoal.model.Postagem;
 import org.generation.blogPessoal.repository.PostagemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,49 +18,55 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
-@RequestMapping ("/postagens")
-@CrossOrigin(origins = "*", allowedHeaders = "*")
+@RequestMapping("/postagens")
+@CrossOrigin(value = "*", allowedHeaders = "*")
 public class PostagemController {
 
 	@Autowired
-	private PostagemRepository repository; 
-	
+	private PostagemRepository repository;
+
 	@GetMapping
-	public ResponseEntity<List<postagem>> GetAll() { 
-		return ResponseEntity.ok(repository.findAll());	
+	public ResponseEntity<List<Postagem>> getAll() {
+		return ResponseEntity.ok(repository.findAll());
 	}
-	
-	
-	@GetMapping ("/{id}")
-	public ResponseEntity<postagem> GetById(@PathVariable long id) {
-		return repository.findById(id)
-				.map(resp -> ResponseEntity.ok(resp))
-				.orElse(ResponseEntity.notFound().build());
+
+	@GetMapping("/{id}")
+	public ResponseEntity<Postagem> getById(@PathVariable long id) {
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
-		    
-	@GetMapping("/titulo/{titulo} ")
-	public ResponseEntity<List<postagem>> GetByTitulo(@PathVariable String titulo) {
+
+	@GetMapping("/titulo/{titulo}")
+	public ResponseEntity<List<Postagem>> getByTitulo(@PathVariable String titulo) {
 		return ResponseEntity.ok(repository.findAllByTituloContainingIgnoreCase(titulo));
 	}
-	
+
 	@PostMapping
-	public ResponseEntity<postagem> post (@RequestBody postagem postagem){ 
-	return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
+	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem postagem) {
+		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
 	}
-	
+
 	@PutMapping
-	public ResponseEntity<postagem> put (@RequestBody postagem postagem){ 
-	return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
+	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem) {
+		return repository.findById(postagem.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
-	
-	@DeleteMapping ("/{id}")
+
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
 	public void delete(@PathVariable long id) {
+		Optional<Postagem> post = repository.findById(id);
+
+		if (post.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
 		repository.deleteById(id);
 	}
-		
-	
+
 }
